requirements.txt
python-telegram-bot==13.15
requests
beautifulsoup4
lxml
